requirements.txt
python-telegram-bot==21.6
nest-asyncio